requirements.txt
cryptography
python-dotenv
django
djangorestframework
django-cors-headers
drf-jwt
pyjwt
requests
